imedit/requirements.txt
Pillow
numpy
opencv-python
scipy